chore(release): Fix release script

+++ scripts/release.sh
@@ -1,12 +1,16 @@
 #! /bin/sh
 
+if test -n "$(git status --porcelain)"; then
+  echo "Your git directory is unclean"
+  exit
+fi
+
 current=`cat VERSION`
 read -p "New version number (current is ${current}): " version
-rm -f VERSION && echo $version > VERSION
-npm run build
 npm version $version
-git commit package.json VERSION dist/ -m "v$version"
-git tag "v$version"
+npm run build
+rm -f VERSION && echo $version > VERSION
+git commit VERSION dist/ --amend -m "v$version"
 git push
 git push --tags
 npm publish
